pyproject.toml
[project]
name = "discord-receipts"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "mit"}
readme = "README.md"
requires-python = "3.11.11"
dependencies = [
    "discord.py-self>=2.1.0",
    "python-escpos>=3.0",
    "Pillow>=10.0.0",
    "requests>=2.31.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
